// frontend/src/api/model/getQuestionTypeOfQuestion.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * TT API
 * Team tuner API documentation
 * OpenAPI spec version: 1.0
 */

export type GetQuestionTypeOfQuestion =
  (typeof GetQuestionTypeOfQuestion)[keyof typeof GetQuestionTypeOfQuestion];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetQuestionTypeOfQuestion = {
  extroversion: 'extroversion',
  agreeableness: 'agreeableness',
  conscientiousness: 'conscientiousness',
  neuroticism: 'neuroticism',
  openness: 'openness',
} as const;