// frontend/src/api/model/outQuestioneItemOut.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * TT API
 * Team tuner API documentation
 * OpenAPI spec version: 1.0
 */
import type { OutQuesitonOutType } from './outQuesitonOutType';

export interface OutQuestioneItemOut {
  answer_id: string;
  answer_value?: number;
  question: string;
  question_id: string;
  question_type: OutQuesitonOutType;
}